+++ Composer/packages/client/src/components/NavLinks/index.tsx
@@ -6,8 +6,9 @@
 import { jsx } from '@emotion/core';
 import React, { Fragment } from 'react';
 import { DefaultButton } from 'office-ui-fabric-react/lib/Button';
+import formatMessage from 'format-message';
 
-import { dialogItemSelected, dialogItemNotSelected } from '../../pages/language-understanding/styles';
+import { dialogItemSelected, dialogItemNotSelected } from './styles';
 interface NavLinksProps {
   navLinks: any[];
   fileId: string;
@@ -28,7 +29,7 @@ export const NavLinks: React.FC<NavLinksProps> = props => {
             }}
             styles={fileId === dialog.id ? dialogItemSelected : dialogItemNotSelected}
             text={dialog.name}
-            ariaLabel={dialog.id + 'language understanding file'}
+            ariaLabel={formatMessage('language understanding file')}
             ariaHidden={false}
           />
         );
